+++ lib/src/widgets/other_widget.dart
@@ -1,0 +1,14 @@
+/*
+* Created By Mirai Devs.
+* On 9/10/2023.
+*/
+import 'package:flutter/material.dart';
+
+class OtherWidget extends StatelessWidget {
+  const OtherWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}